requirements.txt
APScheduler==3.11.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
et_xmlfile==2.0.0
Flask==3.1.0
Flask-APScheduler==1.13.1
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
six==1.17.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
